+++ sql/log_event.cc
@@ -43,8 +43,10 @@
 #include <mysql/psi/mysql_statement.h>
 #include <strfunc.h>
 #include "compat56.h"
-#include "wsrep_mysqld.h"
 #include "sql_insert.h"
+#ifdef WITH_WSREP
+#include "wsrep_mysqld.h"
+#endif /* WITH_WSREP */
 #else
 #include "mysqld_error.h"
 #endif /* MYSQL_CLIENT */
