+++ sql/sql_view.cc
@@ -37,6 +37,7 @@
 #include "sql_derived.h"
 #include "sql_cte.h"    // check_dependencies_in_with_clauses()
 #include "opt_trace.h"
+#include "wsrep_mysqld.h"
 
 #define MD5_BUFF_LENGTH 33
